bot_TG_news/requirements.txt
# Основные зависимости
python-telegram-bot==20.7
beautifulsoup4==4.12.2
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0

# Дополнительные зависимости для улучшенной функциональности
lxml==4.9.3
aiohttp==3.9.1
psutil==5.9.6

# Для разработки и тестирования
pytest==7.4.3
black==23.11.0
flake8==6.1.0
